Fix recommended changes for the PR#1 = minor refactoring
