+++ pta-intelligent-search-harvester/src/test/java/fi/maanmittauslaitos/pta/search/utils/RegionTest.java
@@ -11,11 +11,11 @@ public class RegionTest {
 
     @Test
     public void test_region_intersections() {
-        Region finland = new Region(Arrays.asList(19.450000, 59.780000, 31.610000, 70.120000));
-        Region keskiSuomi = new Region(Arrays.asList(24.010000, 61.450000, 26.780000, 63.610000));
-        Region jklMunicipality = new Region(Arrays.asList(25.260000, 61.840000, 26.050000, 62.430000));
-        Region jkl = new Region(Arrays.asList(25.26078, 61.83952, 26.25, 62.5));
-        Region inari = new Region(Arrays.asList(24.900000, 68.280000, 29.340000, 69.850000));
+        RegionImpl finland = new RegionImpl(Arrays.asList(19.450000, 59.780000, 31.610000, 70.120000));
+        RegionImpl keskiSuomi = new RegionImpl(Arrays.asList(24.010000, 61.450000, 26.780000, 63.610000));
+        RegionImpl jklMunicipality = new RegionImpl(Arrays.asList(25.260000, 61.840000, 26.050000, 62.430000));
+        RegionImpl jkl = new RegionImpl(Arrays.asList(25.26078, 61.83952, 26.25, 62.5));
+        RegionImpl inari = new RegionImpl(Arrays.asList(24.900000, 68.280000, 29.340000, 69.850000));
 
         assertThat(finland.intersects(keskiSuomi)).isTrue();
         assertThat(finland.intersects(jklMunicipality)).isTrue();
